+++ weldx/time.py
@@ -426,7 +426,7 @@ class Time:
         other = Time(other)
         if self.reference_time != other.reference_time:
             return False
-        return np.allclose(self.as_quantity(), other.as_quantity())
+        return np.allclose(self.as_quantity("s").m, other.as_quantity("s").m)
 
     def as_quantity(self, unit: str = "s") -> pint.Quantity:
         """Return the data as `pint.Quantity`.
@@ -453,7 +453,7 @@ class Time:
         if self.is_absolute:
             # store time_ref info
             q.time_ref = self.reference_time  # type: ignore[attr-defined]
-        return q
+        return q  # type: ignore[return-value]
 
     def as_timedelta(self) -> Union[Timedelta, TimedeltaIndex]:
         """Return the data as `pandas.TimedeltaIndex` or `pandas.Timedelta`."""
@@ -574,7 +574,7 @@ class Time:
         `Time.as_quantity`
 
         """
-        return self.as_quantity(unit="s")
+        return self.as_quantity(unit="s")  # type: ignore[return-value]
 
     @property
     def duration(self) -> Time: