+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixes
 
 - fix compatibility with `pint=0.21` \[{pull}`876`\].
+- fix typing compatibility with `pint=0.22` \[{pull}`880`\].
 - add `read_buffer_context` and `write_read_buffer_context` to `weldx.asdf.util`
   to fix tests accessing closed files. {issue}`873` \[{pull}`875`\].
 
@@ -16,6 +17,10 @@
 
 - removed keyword `dummy_inline_arrays` from function `asdf.util.write_buffer`
   as it was only used internally \[{pull}`875`\].
+
+### ASDF
+
+- update `PintQuantityConverter` and `PintUnitConverter` class assignments for `pint=0.22` \[{pull}`880`\].
 
 ## 0.6.6 (19.04.2023)